backend_app/requirements.txt
pyinstaller
pyserial